+++ CHANGELOG.md
@@ -1,6 +1,5 @@
-
 ### Changes 
-* [CHANGE] Delete unnecesary metrics
+* [CHANGE] Add enable-full-metrics flag to enable advanced metrics
 
 ## 0.2.0 / 2019-06-15